Add build matrix to CI. (#3)
This has configurations for Windows + [clang, gcc], but I was only able
to compile successfully on Windows + MSVC. Logs with the full matrix
enabled are here:
https://github.com/iree-org/iree-template-runtime-cmake/actions/runs/6670553422.

We have a similar build matrix upstream in IREE itself, but this is much
simpler and easier to reference. We can point developers at this when
they want to debug their own build environments.